Create SIMPLE_HAND_GESTURE_CONVERTER
This is the Simple hand gesture converter, which converts the hand gestures into texts and displays it. Currently it consists only number, further we can update the  code into a complete hand gesture converter by training them with several data samples. 
Hope you this will give you  a basic idea on how hand gesture converter are. 
This code is fully functional in python which is an advantage to easy understanding and updating.